LDDB has been implemented with 2 way symmetry of ingress/egress coordinates (i.e. approx half the size by only having entries in hash tables for edges that are 0 <= angle < 180 degrees) and 4 way rotational symmetry on map configurations

+++ src/engine/graph/BlockAStar/LDDB/LDDB.java
@@ -18,9 +18,9 @@ public class LDDB implements java.io.Serializable {
 	
 	public double getLength(int mapCode, PairOfCoords p) {
 			if(list.get(mapCode).get(p) == null) {
-				return Double.POSITIVE_INFINITY;
+				return Float.POSITIVE_INFINITY;
 			} else {
-				return list.get(mapCode).get(p).get0().doubleValue();
+				return list.get(mapCode).get(p).get0();
 			} 
 	}
